fix: add `destroy` functionality
